Extract utils from flamechart-view.tsx
